Move BoxSizeParser into DataSource class
This simplifies some dependencies.
