+++ contrib/MBIIDMO.py
@@ -1,6 +1,5 @@
 import numpy
 from nbodykit.plugins import DataSource
-from nbodykit.utils.pluginargparse import BoxSizeParser
 from nbodykit.utils import selectionlanguage
 import os.path
 import logging
@@ -17,7 +16,7 @@ class DataSourcePlugin(DataSource):
         h.add_argument("path", help="path to file")
         h.add_argument("simulation", help="name of simulation", choices=["dmo", "mb2"])
         h.add_argument("type", help="type of objects", choices=["Centrals", "Satellites", "Both"])
-        h.add_argument("BoxSize", type=BoxSizeParser,
+        h.add_argument("BoxSize", type=kls.BoxSizeParser,
             help="the size of the isotropic box, or the sizes of the 3 box dimensions")
             
         h.add_argument("-rsd",
