+++ nbodykit/datasource/PlainText.py
@@ -1,5 +1,4 @@
 from nbodykit.plugins import DataSource
-from nbodykit.utils.pluginargparse import BoxSizeParser
 
 import numpy
 import logging
@@ -64,7 +63,7 @@ class PlainTextDataSource(DataSource):
         h.add_argument("path", help="path to file")
         h.add_argument("names", type=list_str, 
             help="names of columns in file")
-        h.add_argument("BoxSize", type=BoxSizeParser,
+        h.add_argument("BoxSize", type=kls.BoxSizeParser,
             help="the size of the isotropic box, or the sizes of the 3 box dimensions")
         
         h.add_argument("-usecols", type=list_str,
