+++ nbodykit/utils/pluginargparse.py
@@ -169,21 +169,3 @@ class HelpFormatterColon(RawTextHelpFormatter):
         # return the text
         return text
 
-import numpy
-def BoxSizeParser(value):
-    """
-    Parse a string of either a single float, or 
-    a space-separated string of 3 floats, representing 
-    a box size. Designed to be used by the Painter plugins
-    
-    Returns
-    -------
-    BoxSize : array_like
-        an array of size 3 giving the box size in each dimension
-    """
-    boxsize = numpy.empty(3, dtype='f8')
-    sizes = [float(i) for i in value.split()]
-    if len(sizes) == 1: sizes = sizes[0]
-    boxsize[:] = sizes
-    return boxsize
-
